refactor: move instance to runtime_structure module
