+++ src/execution/runtime_structure/export_instances.rs
@@ -1,0 +1,2 @@
+// TODO add ExportInst type. Currently we store exports as a BTreeMap<String, ExternVal> in the
+// ModuleInst, but this is not how the spec does it
