+++ src/execution/runtime_structure/mod.rs
@@ -1,3 +1,11 @@
 pub mod addresses;
+pub mod data_instances;
+pub mod element_instances;
+pub mod export_instances;
+pub mod function_instances;
+pub mod global_instances;
+pub mod memory_instances;
+pub mod module_instances;
+pub mod table_instances;
 pub mod value_stack;
 pub mod values;
